fix: redirect issues live
